Add DevTools in kotlin
