--- redux-java-devtools/src/main/java/com/redux/devtools/Converter.java
@@ -1,8 +1,0 @@
-package com.redux.devtools;
-
-public interface Converter<E> {
-
-    E fromJson(String json);
-
-    String toJson(E element);
-}
